commet

+++ src/main/java/com/example/elearningplatform/course/comment/CommentService.java
@@ -113,7 +113,9 @@ public class CommentService {
                     .orElseThrow(() -> new CustomException("Comment not found", HttpStatus.NOT_FOUND));
             Lesson lesson = commentRepository.findLesson(commentId)
                     .orElseThrow(() -> new CustomException("Lesson not found", HttpStatus.NOT_FOUND));
-            checkCommentAuth(lesson.getId());
+            if (checkCommentAuth(lesson.getId()).equals(false)) {
+                throw new CustomException("Unauthorized", HttpStatus.UNAUTHORIZED);
+            }
             commentRepository.likeComment(tokenUtil.getUserId(), commentId);
 
             comment.incrementNumberOfLikes();
@@ -131,10 +133,11 @@ public class CommentService {
      * @throws Exception ************************************************************************************
      */
 
-    public List<CommentDto> getCommentsByLessonId(Integer lessonId, Integer pageNumber) throws Exception {
-
-        checkCommentAuth(lessonId);
-
+    public Response getCommentsByLessonId(Integer lessonId, Integer pageNumber) throws Exception {
+        try {
+            if (checkCommentAuth(lessonId).equals(false)) {
+                throw new CustomException("Unauthorized", HttpStatus.UNAUTHORIZED);
+            }
             List<Comment> commentes = commentRepository.findByLessonId(lessonId, PageRequest.of(pageNumber, 5));
             List<Comment> likedComments = commentRepository.findLikedCommentsByUserIdAndLesson(tokenUtil.getUserId(),
                     lessonId);
@@ -143,10 +146,14 @@ public class CommentService {
                     comment,
                     likedComments.contains(comment),
                     comment.getUser().getId().equals(tokenUtil.getUserId()))).toList();
-            return commentesDto;
-
+            return new Response(HttpStatus.OK, "Comments fetched successfully", commentesDto);
+        } catch (CustomException e) {
+            return new Response(e.getStatus(), e.getMessage(), null);
+        } catch (Exception e) {
+            return new Response(HttpStatus.INTERNAL_SERVER_ERROR, "Internal Server Error", e.getMessage());
 
     }
+}
 
     public Response removeLikeComment(Integer commentId) {
         try {
@@ -155,7 +162,10 @@ public class CommentService {
 
             Lesson lesson = commentRepository.findLesson(commentId)
                     .orElseThrow(() -> new CustomException("Lesson not found", HttpStatus.NOT_FOUND));
-            checkCommentAuth(lesson.getId());
+
+            if (checkCommentAuth(lesson.getId()).equals(false)) {
+                throw new CustomException("Unauthorized", HttpStatus.UNAUTHORIZED);
+            }
             comment.decrementNumberOfLikes();
             commentRepository.save(comment);
             commentRepository.removeLikeFromComment(tokenUtil.getUserId(), commentId);
@@ -168,11 +178,16 @@ public class CommentService {
         }
     }
     
-    public void checkCommentAuth(Integer lessonId) throws Exception {
+    public Boolean checkCommentAuth(Integer lessonId) {
+        try {
 
         Course course = lessonRepository.findCourseByLessonId(lessonId)
                 .orElseThrow(() -> new CustomException("Course not found", HttpStatus.NOT_FOUND));
-        if (courseService.ckeckCourseSubscribe(course.getId()).equals(false))
-            throw new CustomException("Unauthorized", HttpStatus.BAD_REQUEST);
+        return courseService.ckeckCourseSubscribe(course.getId());
+
+    } catch (Exception e) {
+        return false;
+
+    }
     }
 }
